replace all double with float

+++ apps/main.cpp
@@ -3,7 +3,7 @@
 #include <mps/corr_compressed.h>
 
 #include <array>
-//#include <boost/math/distributions/normal.hpp>
+// #include <boost/math/distributions/normal.hpp>
 #include <cmath>
 #include <iostream>
 // #include "cuPC/cuPC-S.h"
@@ -12,18 +12,18 @@
 
 const int NUMBER_OF_LEVELS = 50;
 
-auto std_normal_qnorm(const double p) -> double
+auto std_normal_qnorm(const float p) -> float
 {
     boost::math::normal dist(0.0, 1.0);
     return quantile(dist, p);
 }
 
-auto threshold_array(const int n, const double alpha) -> std::array<double, NUMBER_OF_LEVELS>
+auto threshold_array(const int n, const float alpha) -> std::array<float, NUMBER_OF_LEVELS>
 {
-    std::array<double, NUMBER_OF_LEVELS> thr{0.0};
+    std::array<float, NUMBER_OF_LEVELS> thr{0.0};
     // my loop range is exclusive of the last i, so I don't subtract one here
     const int n_thr = (NUMBER_OF_LEVELS < (n - 3)) ? NUMBER_OF_LEVELS : n;
-    const double half = 0.5;
+    const float half = 0.5;
     for (size_t i = 0; i < n_thr; i++)
     {
         thr[i] = abs(std_normal_qnorm(half * alpha) / sqrt(n - i - 3));
@@ -35,20 +35,23 @@ auto threshold_array(const int n, const double alpha) -> std::array<double, NUMB
 // {
 //     const int n = 10;
 //     int p = 2;
-//     const double alpha = 0.05;
+//     const float alpha = 0.05;
 //     int max_level = 14;
 //     const size_t sepset_size = p * p * 14;
-//     std::array<double, 4> pmax{0.0};
+//     std::array<float, 4> pmax{0.0};
 //     std::array<int, 4> G = {0, 1, 1, 0};
-//     std::array<double, 4> C = {1.0, 0.4, 0.4, 1.0};
-//     std::array<double, NUMBER_OF_LEVELS> Th = threshold_array(n, alpha);
+//     std::array<float, 4> C = {1.0, 0.4, 0.4, 1.0};
+//     std::array<float, NUMBER_OF_LEVELS> Th = threshold_array(n, alpha);
 //     int sepset[sepset_size];
 //     memset(sepset, 0, sizeof(sepset));
 //     int l = 0;
 //     Skeleton(C.data(), &p, G.data(), Th.data(), &l, &max_level, pmax.data(), sepset);
 // }
 
-auto corr_matrix_size(size_t num_markers) -> size_t { return num_markers * (num_markers - 1) / 2; }
+auto corr_matrix_size(size_t num_markers) -> size_t
+{
+    return num_markers * (num_markers - 1) / 2;
+}
 
 // void call_cu_corr()
 // {
